feat: add custom font selection and fix various launch bugs
