+++ Cargo.toml
@@ -45,6 +45,7 @@ windows = { version = "0.61.3", features = [
     "Win32_Devices_PortableDevices",
     "Win32_NetworkManagement_NetManagement",
     "Win32_UI_Input_KeyboardAndMouse",
+    "Win32_Graphics_DirectWrite",
 ] }
 windows-core = "0.61.2"
 crossbeam-channel = "0.5.15"
