--- src/core/benchmark.rs
@@ -1,209 +1,0 @@
-use crate::core::fs::{parallel_directory_scan, scan_dir_async};
-use crossbeam_channel::{Sender, unbounded};
-use std::path::PathBuf;
-use std::time::Instant;
-
-pub struct BenchmarkResult {
-    pub operation: String,
-    pub duration_ms: u64,
-    pub items_processed: usize,
-    pub memory_usage_mb: f64,
-}
-
-impl BenchmarkResult {
-    pub fn new(operation: String, duration_ms: u64, items_processed: usize) -> Self {
-        Self {
-            operation,
-            duration_ms,
-            items_processed,
-            memory_usage_mb: Self::get_memory_usage(),
-        }
-    }
-
-    fn get_memory_usage() -> f64 {
-        // Simple memory usage estimation
-        // In a real implementation, you'd use psutil or similar
-        0.0 // Placeholder
-    }
-
-    pub fn items_per_second(&self) -> f64 {
-        if self.duration_ms == 0 {
-            0.0
-        } else {
-            self.items_processed as f64 / (self.duration_ms as f64 / 1000.0)
-        }
-    }
-}
-
-pub struct BenchmarkSuite {
-    results: Vec<BenchmarkResult>,
-}
-
-impl BenchmarkSuite {
-    pub fn new() -> Self {
-        Self {
-            results: Vec::new(),
-        }
-    }
-
-    pub fn run_directory_scan_benchmark(&mut self, path: PathBuf) -> &BenchmarkResult {
-        println!("🚀 Benchmarking directory scan: {:?}", path);
-
-        let start = Instant::now();
-        let (tx, rx) = unbounded();
-        let mut count = 0;
-
-        // Start async scan
-        scan_dir_async(path.clone(), tx, true); // Use 24-hour format for benchmarking
-
-        // Count results
-        while let Ok(_) = rx.recv() {
-            count += 1;
-        }
-
-        let duration = start.elapsed();
-        let result = BenchmarkResult::new(
-            format!("Directory Scan ({})", path.display()),
-            duration.as_millis() as u64,
-            count,
-        );
-
-        self.results.push(result);
-        self.results.last().unwrap()
-    }
-
-    pub fn run_folder_size_benchmark(&mut self, path: PathBuf) -> &BenchmarkResult {
-        println!("📁 Benchmarking folder size calculation: {:?}", path);
-
-        let start = Instant::now();
-        let (tx, rx) = unbounded();
-        let mut total_size = 0;
-        let mut completed = false;
-
-        // Start parallel scan
-        parallel_directory_scan(path.clone(), tx);
-
-        // Collect results
-        while let Ok((_, size, done)) = rx.recv() {
-            total_size = size;
-            if done {
-                completed = true;
-                break;
-            }
-        }
-
-        let duration = start.elapsed();
-        let result = BenchmarkResult::new(
-            format!("Folder Size ({})", path.display()),
-            duration.as_millis() as u64,
-            1, // One folder calculated
-        );
-
-        self.results.push(result);
-        self.results.last().unwrap()
-    }
-
-    pub fn run_startup_benchmark(&mut self) -> &BenchmarkResult {
-        println!("⚡ Benchmarking application startup...");
-
-        let start = Instant::now();
-
-        // Simulate startup operations
-        std::thread::sleep(std::time::Duration::from_millis(800)); // Placeholder for actual startup
-
-        let duration = start.elapsed();
-        let result = BenchmarkResult::new(
-            "Application Startup".to_string(),
-            duration.as_millis() as u64,
-            1,
-        );
-
-        self.results.push(result);
-        self.results.last().unwrap()
-    }
-
-    pub fn print_summary(&self) {
-        println!("\n📊 **Benchmark Results**");
-        println!("| Operation | Duration (ms) | Items/s | Memory (MB) |");
-        println!("|-----------|---------------|---------|--------------|");
-
-        for result in &self.results {
-            println!(
-                "| {} | {} | {:.1} | {:.1} |",
-                result.operation,
-                result.duration_ms,
-                result.items_per_second(),
-                result.memory_usage_mb
-            );
-        }
-
-        println!("\n🎯 **Performance Summary**:");
-        if self.results.len() >= 2 {
-            let avg_scan_time: f64 = self
-                .results
-                .iter()
-                .filter(|r| r.operation.contains("Directory Scan"))
-                .map(|r| r.duration_ms as f64)
-                .sum::<f64>()
-                / self
-                    .results
-                    .iter()
-                    .filter(|r| r.operation.contains("Directory Scan"))
-                    .count() as f64;
-
-            println!("- Average directory scan time: {:.1}ms", avg_scan_time);
-            println!("- Total benchmarks run: {}", self.results.len());
-        }
-    }
-
-    pub fn export_markdown_table(&self) -> String {
-        let mut table = String::new();
-        table.push_str("| Operation | EdenExplorer | Windows Explorer | Improvement |\n");
-        table.push_str("|-----------|---------------|------------------|-------------|\n");
-
-        // Add estimated comparisons based on typical performance
-        for result in &self.results {
-            let explorer_estimate = estimate_windows_explorer_time(&result.operation);
-            let improvement = if explorer_estimate > 0 {
-                explorer_estimate as f64 / result.duration_ms as f64
-            } else {
-                1.0
-            };
-
-            table.push_str(&format!(
-                "| {} | ~{}ms | ~{}ms | **{:.1}x faster** |\n",
-                result.operation, result.duration_ms, explorer_estimate, improvement
-            ));
-        }
-
-        table
-    }
-}
-
-fn estimate_windows_explorer_time(operation: &str) -> u64 {
-    // Rough estimates based on typical Windows Explorer performance
-    if operation.contains("Directory Scan") {
-        2500 // ~2.5s for large directories
-    } else if operation.contains("Folder Size") {
-        3200 // ~3.2s for recursive size calculation
-    } else if operation.contains("Startup") {
-        2100 // ~2.1s startup time
-    } else {
-        1000 // Default 1s for other operations
-    }
-}
-
-pub fn run_comprehensive_benchmark(test_path: PathBuf) -> String {
-    let mut suite = BenchmarkSuite::new();
-
-    // Run benchmarks
-    suite.run_startup_benchmark();
-    suite.run_directory_scan_benchmark(test_path.clone());
-    suite.run_folder_size_benchmark(test_path);
-
-    // Print results to console
-    suite.print_summary();
-
-    // Export markdown table
-    suite.export_markdown_table()
-}
