+++ src/core/mod.rs
@@ -1,7 +1,6 @@
-// Expose all core utilities under core::*
-pub mod benchmark;
 pub mod drives;
 pub mod fs;
 pub mod indexer;
 pub mod networkdevices;
 pub mod portable;
+pub mod utils;
