+++ src/main.rs
@@ -2,17 +2,19 @@
 mod core;
 mod gui;
 
-use crate::core::indexer::{WindowSizeMode, load_app_settings, load_windows_size_mode_on_start};
-use crate::gui::windows::windowsoverrides::{get_hwnd_from_cc, set_egui_ctx};
+use crate::core::indexer::{WindowSizeMode, load_windows_size_mode_on_start};
+use crate::core::utils::fonts::apply_custom_font_definitions;
+use crate::gui::windows::windowsoverrides::set_egui_ctx;
 use eframe::{NativeOptions, egui};
 use windows::Win32::System::Com::{COINIT_APARTMENTTHREADED, CoInitializeEx};
 use windows::Win32::UI::WindowsAndMessaging::{GetSystemMetrics, SM_CXSCREEN, SM_CYSCREEN};
 
 fn main() -> eframe::Result<()> {
     unsafe {
-        CoInitializeEx(None, COINIT_APARTMENTTHREADED).unwrap();
+        let _ = CoInitializeEx(None, COINIT_APARTMENTTHREADED);
     }
-    let icon = load_icon().expect("Failed to load icon");
+
+    let icon = load_icon();
     let window_size_mode = load_windows_size_mode_on_start();
     let window_size = match window_size_mode {
         WindowSizeMode::FullScreen => egui::Vec2::new(1920.0, 1080.0),
@@ -39,41 +41,34 @@ fn main() -> eframe::Result<()> {
         options,
         Box::new(|cc| {
             let mut fonts = egui::FontDefinitions::default();
-            egui_phosphor::add_to_fonts(&mut fonts, egui_phosphor::Variant::Regular);
-            // Register fill variant under a separate family for selective use (e.g., filled star).
-            fonts.font_data.insert(
-                "phosphor_fill".to_owned(),
-                egui_phosphor::Variant::Fill.font_data().into(),
-            );
-            fonts.families.insert(
-                egui::FontFamily::Name("phosphor_fill".into()),
-                vec!["phosphor_fill".to_owned()],
-            );
-            fonts.font_data.insert(
-                "japanese_font".to_owned(),
-                egui::FontData::from_static(include_bytes!("assets/NotoSansJP-Regular.ttf")).into(),
-            );
-            for family in &mut fonts.families.values_mut() {
-                family.insert(0, "japanese_font".to_owned());
-            }
+
+            apply_custom_font_definitions(&mut fonts);
 
             cc.egui_ctx.set_fonts(fonts);
-
-            let hwnd = get_hwnd_from_cc(cc);
             set_egui_ctx(&cc.egui_ctx);
-            Ok(Box::new(gui::MainWindow::new(hwnd)))
+
+            Ok(Box::new(gui::MainWindow::new()))
         }),
     )
 }
 
 static ICON_BYTES: &[u8] = include_bytes!("assets/icon.ico");
 
-fn load_icon() -> Option<egui::IconData> {
-    let image = image::load_from_memory(ICON_BYTES).ok()?.into_rgba8();
-    let (width, height) = image.dimensions();
-    Some(egui::IconData {
-        rgba: image.into_raw(),
-        width,
-        height,
-    })
+fn load_icon() -> egui::IconData {
+    match image::load_from_memory(ICON_BYTES) {
+        Ok(img) => {
+            let rgba = img.into_rgba8();
+            let (width, height) = rgba.dimensions();
+            egui::IconData {
+                rgba: rgba.into_raw(),
+                width,
+                height,
+            }
+        }
+        Err(_) => egui::IconData {
+            rgba: vec![0u8; 64 * 64 * 4],
+            width: 64,
+            height: 64,
+        },
+    }
 }
